Properly fix the progress indicators.

+++ src/main/java/epicsquid/roots/integration/top/TOPPlugin.java
@@ -8,10 +8,8 @@ import epicsquid.roots.tileentity.TileEntityPyre;
 import mcjty.theoneprobe.api.*;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
-import net.minecraft.client.resources.I18n;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.World;
 
@@ -46,7 +44,7 @@ public class TOPPlugin implements Function<ITheOneProbe, Void>, IProbeInfoProvid
         } else {
           return;
         }
-        double totalSeconds = (duration - (duration - te.getBurnTime()) / 20.0);
+        int totalSeconds = (int) ((duration - (duration - te.getBurnTime())) / 20.0);
         probeInfo.text("{*roots.hud.top.pyre.progress*}" + totalSeconds + " {*roots.hud.top.pyre.progress_info*} ");
       }
     }
